Revert back to nearly empty repo
Something was borked with the Dioxus setup and it refused to build.
Unfortunately the build error was not at all helpful to me in figuring
out what the problem is. I could maybe hunt it down and figure it out,
but first I want to try a slightly different way of doing the setup.